ci: add commit linting (#1)
Motivation
==========

To make history easier to read, commit messages should follow the
convention.

Linter
==========

The commitlint CLI is used to validate commit messages against a
specific format.

The format chosen is _@commitlint/config-conventional_, which sets the
rules for Conventional
Commits.

Non-main branches
==========

Commits in non-main branches should follow the convention to make it
easier for reviewers to read
pull request histories.

Validation can be done using GitHub Actions or Git Hooks. Validation
with Git Hooks is faster than
GitHub Actions because you don't have to spend time pushing and
installing dependencies, and the
time for the rest of the verification steps should be about the same. I
didn't find any other
significant differences. That's why I chose Git Hooks.

To make it easier to work with GitHook, I installed _husky_.

We need to check for commit messages that are on the pushing branch but
that are not on the remote
branches. The _scripts/lint-commit-messages.mjs_ script does this.

The script is written for executing via _zx_ package like other scripts
in the repository. To
access some of the features from the new API, I updated the _zx_
package.

Main branch
==========

Commits can only get into the main branch through a pull request.
Therefore, it is enough for me
to check only the pull request.

The only pull-request merge strategy available is the squash merge. The
commit message in the main
branch is formed from the title and description of the pull request.
Therefore, I only need to
check the title and description of the pull request.

The _.github/workflows/lint-pull-message.yml_ workflow does this.