feat: add component overlaying image rendering (#4)
This commit will add the feature to add an overlaying image for a component.

An overlaying image is displayed in the same place as the component, that is, the upper left corner
of an overlaying image coincides with the upper left corner of the rectangle in which a component
fits. An overlaying image is displayed with some transparency and inverted colors. An overlaying
image does not respond to pointer events.

To add an overlaying image for the story, you need to set the pixelPerfect.overlaySrc parameter,
in which to pass the path to an image. It can accept any string that the src attribute of the
img tag accepts.