docs: add overlays for other demo stories (#8)
This commit will add overlays for other demo stories that have been in the project since it was
initialized.

Demo stories for which overlays will be added:
- Header;
- Page.

An overlay image has already been added to the Button story.

This commit will also change the default viewport sizes. The viewport sizes will be fixed. This is
necessary so that other people with different screen resolutions can see how the addon works.