docs: remove demo `myAddonParameter` parameter from button stories (#14)
This commit will remove the use of the demo parameter from the stories of the Button component that
has been in the repository since it was initialized. The functionality of this parameter has
already been removed, but an example of its use remains. This commit will remove exactly the use case
for this parameter.