feat: add overlay opacity customization (#17)
This commit will add the feature of setting the opacity of the overlay.

The opacity of the overlay can be adjusted via the story parameter: `pixelPerfect.overlay.opacity`.
Also, the opacity of the overlay can be adjusted through the corresponding control on the
"Pixel Perfect" panel.

If you set the opacity through the control, then the changes will be reset after the page is
reloaded.

The settings set through the control override the settings set through the parameter. If you do not
specify opacity in the parameters or through the control, then the default opacity settings will be
used.

For the control UI, the `@mui/material` library is used. The color of the control will match the
theme of the storybook.

This commit will also change how the overlay is rendered. Previously, the overlay element would be
deleted and recreated with new settings when those settings were changed. Because of this, it was
possible to see the moment between deletion and re-creation, when the overlay is not displayed,
when changing settings through the control, which could interfere with adjusting the overlay
settings. Now the overlay element is simply updated if the overlay element already exists.