feat: add feature to toggle color inversion of the overlay (#18)
This commit will add a feature to toggle color inversion of the overlay so that you can better
customize the overlay for different situations.

The color inversion of the overlay can be set via the story parameter:
`pixelPerfect.overlay.colorInversion`. Also, it can be set via the corresponding control on the
addon panel in Storybook.

If you set color inversion via the panel, then the changes will be reset after the page is
reloaded.

The settings set via the control override the settings set via the parameter. If you do not set
color inversion in the parameters or via the control, then the default color inversion settings
will be used.