feat: add feature to reset overlay options (#19)
This commit will add the ability to reset overlay options set via the addon panel to revert the
overlay to its previous state after experimenting with its options. To do this, reset buttons will
be added to the addon panel.

There are reset buttons for each option, as well as one reset button for all options at once.

When the current values of the options differ from the initial values, then the buttons
corresponding to the reset buttons are highlighted. This is necessary to understand which options
have been changed as a result of customization.

The `@mui/icons-material` package has been installed for reset icon.